BACnet.Protocol.LogStatus (bacstack v0.1.0)

Types

t()

@type t() :: %BACnet.Protocol.LogStatus{
  buffer_purged: boolean(),
  log_disabled: boolean(),
  log_interrupted: boolean()
}

Functions

encode(flags, opts \\ [])

@spec encode(t(), Keyword.t()) ::
  {:ok, BACnet.Protocol.ApplicationTags.encoding_list()} | {:error, term()}

Encodes a BACnet log status into application tags encoding.

from_bitstring(bitstring)

@spec from_bitstring({boolean(), boolean()} | {boolean(), boolean(), boolean()}) ::
  t()

Creates from an application tag bitstring a log status.

parse(tags)

@spec parse(BACnet.Protocol.ApplicationTags.encoding_list()) ::
  {:ok, {t(), rest :: BACnet.Protocol.ApplicationTags.encoding_list()}}
  | {:error, term()}

Parses a BACnet log status from application tags encoding.

to_bitstring(t)

@spec to_bitstring(t()) :: BACnet.Protocol.ApplicationTags.primitive_encoding()

Creates an application tag bitstring from a log status.

valid?(t)

@spec valid?(t()) :: boolean()

Validates whether the given log status is in form valid.

It only validates the struct is valid as per type specification.