BACnet.Protocol.ObjectTypes.Command (bacstack v0.1.0)

The Command object provides a single-point "macro" mechanism: writing an integer action code (present_value) causes the object to execute a predefined list of write operations on other object properties, all as a single atomic action from the network's point of view. This is the standard way to implement "go to occupied mode", "night setback", "fire alarm response", or any other coordinated multi-point state change.

Each action in action (a list of BACnet.Protocol.ActionCommand) specifies a target object/property, a value to write (which may be a priority or relinquish), and optional conditions. The action_text array gives human-readable names for each action code.

Object Description (ASHRAE 135)

The Command object type defines a standardized object whose properties represent the externally visible characteristics of a multi-action command procedure.

Behaviour and Operation

Command objects are active "macro" executors. Writing an integer (1..N) to the present_value (the action code) causes the object to iterate over the corresponding entry in the action array (a list of BACnet.Protocol.ActionList) and perform the specified writes to other objects' properties (possibly on remote devices) as a single logical operation.

Developer Implementation Notes (geared to device server / application authors)

The generated code handles storage + basic mechanics (validation, implicit_relationships, readonly annotations as hints to your server, etc.). You must drive "special" live properties and side effects yourself, analogous to maintaining present_value on inputs via update_property/3 (never direct mutation). Read notes below + generated tables for details.

Special / live properties and expected developer behaviour

present_value (non_neg_integer): The action code (1-based index into action). Dev must: This is the trigger, not a normal value.

action, action_text: The "program" of the macro.

all_writes_successful: Indicates last execution status? Dev must: Set/clear based on your execution results.

The Command object is the standard way to expose a "canned sequence of writes" as a single network-visible action. It is deliberately not a priority-array commandable in the usual sense; the act of writing a new action code is the trigger.

Execution is your responsibility: When a client (or your own internal code) successfully does update_property(cmd_obj, :present_value, action_code) you must:

Look up the corresponding ActionList in action.

For each ActionCommand in that list, perform the write it describes:

target object + property (can be a remote DeviceObjectPropertyRef)

value (an Encoding or primitive)

optional priority (for commandable targets) or the special "relinquish" sentinel

You may abort early on error if the quit_on_failure property is set on an action command.

Because the writes can target remote devices you will be issuing real WriteProperty / WritePropertyMultiple service requests (or using your client's write helpers).

action_text: A parallel array of human strings so that an operator workstation can present a nice menu: "1 - Go to Night Setback", "2 - Fire Mode", etc. You populate it at creation time (or allow it to be written later).

Atomicity considerations for developers: The BACnet spec says the writes appear atomic to the network. In practice you will do a series of individual writes. If you want true rollback you need to remember the previous values of the targets (which may require reading them first at the right priorities) and be prepared to write the old values back on failure.

Writing the action list from the wire: Because action (a BACnetArray of ActionList) and action_text are normal writable properties, a sufficiently privileged client can completely reprogram what action N does.

Command objects are one of the few places where a single network write can cause a cascade of other writes - design your execution engine with care around priority, failure handling, and logging so that operators can debug "why did the lights and the AHU both change at 17:00?".

Examples

Creating a Command object:

iex> {:ok, cmd} = BACnet.Protocol.ObjectTypes.Command.create(1300, "ZoneControl", %{}); cmd.object_name
"ZoneControl"

See Also

BACnet.Protocol.ActionCommand

BACnet.Protocol.ActionList

The following part has been automatically generated.

Click to expand
This module defines a BACnet object of the type `command`. The following properties are defined: | Property | Revision | Required | Readonly | Protected | Intrinsic | |----------|----------|----------|----------|-----------|-----------| | action | | X | X | | | | action_text | | | | | | | all_writes_successful | | X | X | | | | description | | | | | | | in_process | | X | | | | | object_instance | | X | X | | | | object_name | | X | X | | | | present_value | | X | | | | | profile_name | | | | | | The following properties have additional semantics: | Property | Has Default | Has Init | Implicit Relationships | Validators | Annotations | |----------|-------------|----------|------------------------|------------|-------------| | action | X | | | | | | action_text | | | | Fun | | | present_value | X | | | | | The following table shows the default values and/or init functions: | Property | Default Value | Init Function | |----------|---------------|---------------| | action | `%BACnet.Protocol.BACnetArray{...}` | | | present_value | `0` | |

Types

common_object_opts()

@type common_object_opts() ::
  {:allow_unknown_properties, boolean()}
  | {:ignore_unknown_properties, boolean()}
  | {:revision, BACnet.Protocol.Constants.protocol_revision()}

Common object options for creation - all are optional.

allow_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties (numeric identifiers usually means we dont know them) are accepted and put into a separate map. This does mean we can not validate or write them. Types of the values can be anything at this point. While you can read unknown properties with atom or integer as property identifier, you can only remove numeric unknown property identifiers from an object. Property identifiers of type atom are only accepted, if it is a remote object (object implementation is only enforced if it is a local object). Numeric property identifiers are accepted regardless of remote object or not. For remote objects, this means you have to write "raw values" (usually Encoding structs).

ignore_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties get ignored, as if they were not specified.

revision - The BACnet protocol revision to check required properties against. Optional properties are regardless of revision available. See BACnet.Protocol.Constants.protocol_revision/0 for the available revisions.

object_opts()

@type object_opts() :: common_object_opts()

Options accepted when creating or configuring a Command object.

property_name()

@type property_name() ::
  :action
  | :action_text
  | :all_writes_successful
  | :description
  | :in_process
  | :object_instance
  | :object_name
  | :present_value
  | :profile_name

Available property names for this object.

property_update_error()

@type property_update_error() ::
  {:error,
   {error :: atom(),
    property :: BACnet.Protocol.Constants.property_identifier()}}

The structure for property errors.

t()

@type t() :: %BACnet.Protocol.ObjectTypes.Command{
  _metadata: internal_metadata(),
  _unknown_properties: %{
    optional(atom() | non_neg_integer()) =>
      term()
      | BACnet.Protocol.ApplicationTags.Encoding.t()
      | [BACnet.Protocol.ApplicationTags.Encoding.t()]
  },
  action: BACnet.Protocol.BACnetArray.t(BACnet.Protocol.ActionList.t()),
  action_text: BACnet.Protocol.BACnetArray.t(String.t()) | nil,
  all_writes_successful: boolean(),
  description: String.t() | nil,
  in_process: boolean(),
  object_instance: non_neg_integer(),
  object_name: String.t(),
  present_value: non_neg_integer(),
  profile_name: String.t() | nil
}

Represents a Command object. All keys should be treated as read-only, all updates should go only through update_property/3.

Functions

add_property(object, property, value)

@spec add_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Adds an optional property to an object. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically added and instead the object must be newly created using create/4.

create(instance_number, object_name, properties \\ %{}, opts \\ [])

@spec create(
  non_neg_integer(),
  String.t(),
  %{optional(property_name() | atom() | non_neg_integer()) => term()},
  [object_opts() | internal_metadata()]
) :: {:ok, t()} | property_update_error()

Creates a new object struct with the defined properties. Optional properties are not created when not given, only required, given and dependency properties are created. Properties with a value of nil are ignored.

Only properties that are required for specific services (i.e. Intrinsic Reporting) are automatically created.

get_all_properties()

@spec get_all_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of all properties this object supports.

get_annotation(name)

@spec get_annotation(property_name()) :: [term()]

Auto generated function to get the annotations for the given property name.

get_annotations()

@spec get_annotations() :: [{name :: property_name(), values :: [term()]}]

Auto generated function to get the list of annotations for each property.

get_cov_properties()

@spec get_cov_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of properties used for COV reporting.

get_intrinsic_properties()

@spec get_intrinsic_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of intrinsic properties.

get_object_identifier(object)

@spec get_object_identifier(t()) :: BACnet.Protocol.ObjectIdentifier.t()

Get the BACnet object identifier.

get_optional_properties()

@spec get_optional_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of optional properties.

get_properties(object)

@spec get_properties(t()) :: [BACnet.Protocol.Constants.property_identifier()]

Get the list of properties the object has.

get_properties_type_map()

@spec get_properties_type_map() :: map()

Auto generated function to get a map of property name to type.

get_property(object, property)

@spec get_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, term()} | property_update_error()

Get a property's value from an object.

get_protected_properties()

@spec get_protected_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of protected properties.

Protected is an annotation and the object modules prevent writing to this property directly in code. The protected properties are either written on creation or updated automatically depending on other properties being written to. Some properties are only written once at creation and never updated.

get_readonly_properties()

@spec get_readonly_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of readonly properties.

Readonly is only an annotation that the property should be write protected on the BACnet side, there is no actual write protection in the object. This is a hint to the device server. If you need actual write protection, see protected.

get_required_properties()

@spec get_required_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of required properties.

has_property?(object, property)

@spec has_property?(t(), BACnet.Protocol.Constants.property_identifier()) :: boolean()

Checks if the given object has the given property.

See BACnet.Protocol.ObjectsUtility.has_property?/2 for implementation details.

property_writable?(object, property)

@spec property_writable?(t(), BACnet.Protocol.Constants.property_identifier()) ::
  boolean()

Checks if the given property is writable.

Check BACnet.Protocol.ObjectsUtility.property_writable?/2 for a basic run-down.

remove_property(object, property)

@spec remove_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, t()} | property_update_error()

Removes an optional property from an object. This function is idempotent. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically removed and instead the object must be newly created using create/4. Required properties can not be removed.

supports_intrinsic()

@spec supports_intrinsic() :: boolean()

Auto generated function to check whether the object type supports intrinsic reporting.

update_property(object, property, value)

@spec update_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Updates a property of an object.