BACnet.Protocol.ObjectTypes.DateTimeValue (bacstack v0.1.0)

The Date Time Value object holds a single concrete date-and-time value. It is the non-recurring counterpart to BACnet.Protocol.ObjectTypes.DateTimePatternValue and is typically used for absolute timestamps such as "last maintenance performed", "contract expires", "next scheduled run", or any other point-in-time configuration or status datum that must be visible on the network.

The value can be commandable via priority array, allowing operators or automated procedures to set or override the timestamp. It is a simple data container; the meaning of the timestamp is a local matter or defined by the objects that read it.

Object Description (ASHRAE 135)

The DateTime Value object type defines a standardized object whose properties represent the externally visible characteristics of a named data value in a BACnet device.

A DateTime Value object is used to represent a single moment in time.

Behaviour and Operation

Date Time Value objects hold a single concrete timestamp. They are used for deadlines, last-run times, configuration points, etc. The timestamp can be read and (if not commandable) written directly by clients or the local application.

When made commandable via a priority array the effective timestamp is taken from the priority mechanism. The object performs no automatic time-based action; any behaviour (e.g. "if now > this value then ...") is implemented in the objects or logic that read it.

Developer Implementation Notes (geared to device server / application authors)

The generated code handles storage + basic mechanics (validation, implicit_relationships, readonly annotations as hints to your server, etc.). You must drive "special" live properties and side effects yourself, analogous to maintaining present_value on inputs via update_property/3 (never direct mutation). Read notes below + generated tables for details.

Special / live properties and expected developer behaviour

present_value: The timestamp value. Dev must: Your app/config/time logic writes it (directly if not commandable; via set_priority/3 if commandable). Consumers read it for deadlines etc. The object does no auto-updating or time keeping.

priority_array, relinquish_default: For commandable timestamp override. Dev must: Use priority mechanism for multiple sources to "set" the time value.

status_flags, out_of_service, reliability: Dev must: out_of_service allows forcing a test timestamp if not commandable. Reliability if the source of this timestamp value is unhealthy. in_alarm/fault/out_of_service bits are auto-updated by the object (overridden local matter).

is_utc: Whether the time is UTC. Dev must: Set/maintain correctly when writing the value.

Commandability and Priority Arrays

Value objects can have a priority_array (making them commandable).

Examples

Creating a DateTime Value:

iex> {:ok, dtv} = BACnet.Protocol.ObjectTypes.DateTimeValue.create(90, "Event", %{present_value: %BACnet.Protocol.BACnetDateTime{date: %BACnet.Protocol.BACnetDate{year: :unspecified, month: :unspecified, day: :unspecified, weekday: :unspecified}, time: %BACnet.Protocol.BACnetTime{hour: :unspecified, minute: :unspecified, second: :unspecified, hundredth: :unspecified}}}); dtv.object_name
"Event"

See Also

BACnet.Protocol.BACnetDateTime

BACnet.Protocol.ObjectTypes.DateTimePatternValue

The following part has been automatically generated.

Click to expand
This module defines a BACnet object of the type `datetime_value`. The following properties are defined: | Property | Revision | Required | Readonly | Protected | Intrinsic | |----------|----------|----------|----------|-----------|-----------| | description | | | | | | | event_state | | | | | | | is_utc | | | | | | | object_instance | | X | X | | | | object_name | | X | X | | | | out_of_service | | | | | | | present_value | | X | | | | | priority_array | | | X | | | | profile_name | | | | | | | reliability | | | | | | | reliability_evaluation_inhibit | | | | | | | relinquish_default | | | | | | | status_flags | | X | X | | | The following properties have additional semantics: | Property | Has Default | Has Init | Implicit Relationships | Validators | Annotations | |----------|-------------|----------|------------------------|------------|-------------| | present_value | | | | | `readonly_when: {:out_of_service, false}` | | reliability | | | reliability_evaluation_inhibit | | | The following table shows the default values and/or init functions: | Property | Default Value | Init Function | |----------|---------------|---------------|

Types

common_object_opts()

@type common_object_opts() ::
  {:allow_unknown_properties, boolean()}
  | {:ignore_unknown_properties, boolean()}
  | {:revision, BACnet.Protocol.Constants.protocol_revision()}

Common object options for creation - all are optional.

allow_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties (numeric identifiers usually means we dont know them) are accepted and put into a separate map. This does mean we can not validate or write them. Types of the values can be anything at this point. While you can read unknown properties with atom or integer as property identifier, you can only remove numeric unknown property identifiers from an object. Property identifiers of type atom are only accepted, if it is a remote object (object implementation is only enforced if it is a local object). Numeric property identifiers are accepted regardless of remote object or not. For remote objects, this means you have to write "raw values" (usually Encoding structs).

ignore_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties get ignored, as if they were not specified.

revision - The BACnet protocol revision to check required properties against. Optional properties are regardless of revision available. See BACnet.Protocol.Constants.protocol_revision/0 for the available revisions.

object_opts()

@type object_opts() :: common_object_opts()

Options accepted when creating or configuring a DateTime Value object.

property_name()

@type property_name() ::
  :description
  | :event_state
  | :is_utc
  | :object_instance
  | :object_name
  | :out_of_service
  | :present_value
  | :priority_array
  | :profile_name
  | :reliability
  | :reliability_evaluation_inhibit
  | :relinquish_default
  | :status_flags

Available property names for this object.

property_update_error()

@type property_update_error() ::
  {:error,
   {error :: atom(),
    property :: BACnet.Protocol.Constants.property_identifier()}}

The structure for property errors.

t()

@type t() :: %BACnet.Protocol.ObjectTypes.DateTimeValue{
  _metadata: internal_metadata(),
  _unknown_properties: %{
    optional(atom() | non_neg_integer()) =>
      term()
      | BACnet.Protocol.ApplicationTags.Encoding.t()
      | [BACnet.Protocol.ApplicationTags.Encoding.t()]
  },
  description: String.t() | nil,
  event_state: BACnet.Protocol.Constants.event_state() | nil,
  is_utc: boolean() | nil,
  object_instance: non_neg_integer(),
  object_name: String.t(),
  out_of_service: boolean() | nil,
  present_value: BACnet.Protocol.BACnetDateTime.t(),
  priority_array:
    BACnet.Protocol.PriorityArray.t(BACnet.Protocol.BACnetDateTime.t()) | nil,
  profile_name: String.t() | nil,
  reliability: BACnet.Protocol.Constants.reliability() | nil,
  reliability_evaluation_inhibit: boolean() | nil,
  relinquish_default: BACnet.Protocol.BACnetDateTime.t() | nil,
  status_flags: BACnet.Protocol.StatusFlags.t()
}

Represents a DateTime Value object. All keys should be treated as read-only, all updates should go only through update_property/3.

Properties which are for Intrinsic Reporting are nil, if disabled. If Intrinsic Reporting is enabled on the object, then the properties can not be nil.

For commandable objects (objects with a priority array), the present value property is protected.

Functions

add_property(object, property, value)

@spec add_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Adds an optional property to an object. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically added and instead the object must be newly created using create/4.

create(instance_number, object_name, properties \\ %{}, opts \\ [])

@spec create(
  non_neg_integer(),
  String.t(),
  %{optional(property_name() | atom() | non_neg_integer()) => term()},
  [object_opts() | internal_metadata()]
) :: {:ok, t()} | property_update_error()

Creates a new object struct with the defined properties. Optional properties are not created when not given, only required, given and dependency properties are created. Properties with a value of nil are ignored.

Only properties that are required for specific services (i.e. Intrinsic Reporting) are automatically created.

get_all_properties()

@spec get_all_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of all properties this object supports.

get_annotation(name)

@spec get_annotation(property_name()) :: [term()]

Auto generated function to get the annotations for the given property name.

get_annotations()

@spec get_annotations() :: [{name :: property_name(), values :: [term()]}]

Auto generated function to get the list of annotations for each property.

get_cov_properties()

@spec get_cov_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of properties used for COV reporting.

get_intrinsic_properties()

@spec get_intrinsic_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of intrinsic properties.

get_object_identifier(object)

@spec get_object_identifier(t()) :: BACnet.Protocol.ObjectIdentifier.t()

Get the BACnet object identifier.

get_optional_properties()

@spec get_optional_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of optional properties.

get_priority_value(object)

@spec get_priority_value(t()) ::
  {priority :: 1..16, value :: BACnet.Protocol.BACnetDateTime.t()} | nil

Get the active priority value from the priority array, or nil.

get_properties(object)

@spec get_properties(t()) :: [BACnet.Protocol.Constants.property_identifier()]

Get the list of properties the object has.

get_properties_type_map()

@spec get_properties_type_map() :: map()

Auto generated function to get a map of property name to type.

get_property(object, property)

@spec get_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, term()} | property_update_error()

Get a property's value from an object.

get_protected_properties()

@spec get_protected_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of protected properties.

Protected is an annotation and the object modules prevent writing to this property directly in code. The protected properties are either written on creation or updated automatically depending on other properties being written to. Some properties are only written once at creation and never updated.

get_readonly_properties()

@spec get_readonly_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of readonly properties.

Readonly is only an annotation that the property should be write protected on the BACnet side, there is no actual write protection in the object. This is a hint to the device server. If you need actual write protection, see protected.

get_required_properties()

@spec get_required_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of required properties.

has_priority_array?(object)

@spec has_priority_array?(t()) :: boolean()

Checks whether the given object has a priority array.

has_property?(object, property)

@spec has_property?(t(), BACnet.Protocol.Constants.property_identifier()) :: boolean()

Checks if the given object has the given property.

See BACnet.Protocol.ObjectsUtility.has_property?/2 for implementation details.

property_writable?(object, property)

@spec property_writable?(t(), BACnet.Protocol.Constants.property_identifier()) ::
  boolean()

Checks if the given property is writable.

Check BACnet.Protocol.ObjectsUtility.property_writable?/2 for a basic run-down.

remove_property(object, property)

@spec remove_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, t()} | property_update_error()

Removes an optional property from an object. This function is idempotent. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically removed and instead the object must be newly created using create/4. Required properties can not be removed.

set_priority(object, priority, value)

@spec set_priority(t(), 1..16, BACnet.Protocol.BACnetDateTime.t() | nil) ::
  {:ok, t()} | property_update_error()

Sets the given priority in the priority array of an object. This function also updates the present value.

supports_intrinsic()

@spec supports_intrinsic() :: boolean()

Auto generated function to check whether the object type supports intrinsic reporting.

update_property(object, property, value)

@spec update_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Updates a property of an object. To update the priority array, use set_priority/3 instead.