BACnet.Protocol.ObjectTypes.EventLog (bacstack v0.1.0)

The Event Log object captures BACnet.Protocol.Services.ConfirmedEventNotification and BACnet.Protocol.Services.UnconfirmedEventNotification into a timestamped buffer, so that alarm and event history can be retrieved later with the BACnet.Protocol.Services.ReadRange service. It is the BACnet-native replacement for a local alarm printer or syslog for events.

Records contain the full event information (source object, event type, message text, priority, etc.). The log can be configured as a fixed-size circular buffer or a growing one (until stopped or full). It supports a time window (start_time / stop_time), enable/disable, and can itself generate a BUFFER_READY intrinsic event (when intrinsic_reporting: true) when the log has new data worth harvesting.

Object Description (ASHRAE 135)

An Event Log object records event notifications with timestamps and other pertinent data in an internal buffer for subsequent retrieval.

Behaviour and Operation

Event Log objects are buffers that receive copies of event notifications (both confirmed and unconfirmed). The conditions by which events are forwarded to a particular Event Log is not specified and is a local matter. The local application is responsible for appending a record (containing timestamp, source, event type, message text, etc.) to the log buffer whenever an event notification is sent or received that matches the log's configuration.

Records are retrieved by clients using the ReadRange service. The log can be circular or append-only until full/stopped. start_time / stop_time and the enable flag control when logging is active. When intrinsic reporting is enabled the log itself can emit a BUFFER_READY event to notify a workstation that new records are available for collection.

Developer Implementation Notes (geared to device server / application authors)

The generated code handles storage + basic mechanics (validation, implicit_relationships, readonly annotations as hints to your server, etc.). You must drive "special" live properties and side effects yourself, via update_property/3 (never direct mutation). Read notes below + generated tables for details.

Special / live properties and expected developer behaviour

Similar to TrendLog but for event notifications (ConfirmedEventNotification etc).

log_buffer: History of event notifications. Dev must: Your event notification handler (when you emit or receive notifications that should be logged) must construct EventLogRecord (with timestamp, the notification params, status) and append via update on the buffer (respect buffer_size, stop_when_full, enable, start/stop times).

record_count, total_record_count: Dev must: Manage counts on appends; for BUFFER_READY when intrinsic, track records since last notify and emit when threshold crossed.

start_time/stop_time, enable, stop_when_full, buffer_size: Config for logging window. Dev must: Your logger respects the times/enable when deciding to log a notification. buffer_size writes restricted when enabled.

status_flags, reliability. Dev must: Update on problems logging or with event sources. Note in_alarm/ fault/out_of_service bits of status_flags are automatically managed by the object.

The object is passive storage + config; you drive the logging of events into it.

Intrinsic Reporting

When intrinsic_reporting: true is passed to create/4, additional event reporting properties become active for the log itself.

Examples

Creating an Event Log:

iex> {:ok, el} = BACnet.Protocol.ObjectTypes.EventLog.create(800, "EventHistory", %{buffer_size: 100}); el.object_name
"EventHistory"

With intrinsic reporting:

iex> {:ok, el} = BACnet.Protocol.ObjectTypes.EventLog.create(801, "Logged", %{buffer_size: 50}, intrinsic_reporting: true); el.object_name
"Logged"

See Also

BACnet.Protocol.EventAlgorithms.BufferReady

BACnet.Stack.TrendLogger

The following part has been automatically generated.

Click to expand
This module defines a BACnet object of the type `event_log`. The following properties are defined: | Property | Revision | Required | Readonly | Protected | Intrinsic | |----------|----------|----------|----------|-----------|-----------| | acked_transitions | | | X | | X | | buffer_size | | X | X | | | | description | | | | | | | enable | | X | | | | | event_algorithm_inhibit | | | | | X | | event_algorithm_inhibit_ref | | | | | X | | event_detection_enable | | | | | X | | event_enable | | | | | X | | event_message_texts | | | X | | X | | event_message_texts_config | | | | | X | | event_state | | X | | | | | event_timestamps | | | X | | X | | last_notify_record | | | | | X | | limit_enable | | | | | X | | log_buffer | | X | | | | | notification_class | | | | | X | | notification_threshold | | | | | X | | notify_type | | | | | X | | object_instance | | X | X | | | | object_name | | X | X | | | | profile_location | 19 | | | | | | profile_name | | | | | | | record_count | | X | | | | | records_since_notification | | | | | X | | reliability | | | | | | | reliability_evaluation_inhibit | | | | | | | start_time | | | | | | | status_flags | | X | X | | | | stop_time | | | | | | | stop_when_full | | X | | | | | tags | 19 | | | | | | time_delay | | | | | X | | time_delay_normal | | | | | X | | total_record_count | | X | X | | | The following properties have additional semantics: | Property | Has Default | Has Init | Implicit Relationships | Validators | Annotations | |----------|-------------|----------|------------------------|------------|-------------| | buffer_size | | | | Fun/Type | | | enable | X | | | | | | event_algorithm_inhibit_ref | | | event_algorithm_inhibit | | | | event_state | X | | | | | | last_notify_record | X | | | Type | | | log_buffer | X | | | Fun | | | notification_threshold | X | | | Type | | | profile_location | | | | Fun | `revision: 19` | | record_count | X | | | Type | | | records_since_notification | X | | | Type | | | reliability | | | reliability_evaluation_inhibit | | | | start_time | | | stop_time | | | | tags | | | | | `revision: 19` | | total_record_count | X | | | Type | | The following table shows the default values and/or init functions: | Property | Default Value | Init Function | |----------|---------------|---------------| | enable | `true` | | | event_state | `:normal` | | | last_notify_record | `0` | | | log_buffer | `[]` | | | notification_threshold | `0` | | | record_count | `0` | | | records_since_notification | `0` | | | total_record_count | `0` | |

Types

common_object_opts()

@type common_object_opts() ::
  {:allow_unknown_properties, boolean()}
  | {:ignore_unknown_properties, boolean()}
  | {:revision, BACnet.Protocol.Constants.protocol_revision()}
  | {:skip_property_validation_remote_object, boolean() | :value}

Common object options for creation - all are optional.

allow_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties (numeric identifiers usually means we dont know them) are accepted and put into a separate map. This does mean we can not validate or write them. Types of the values can be anything at this point. While you can read unknown properties with atom or integer as property identifier, you can only remove numeric unknown property identifiers from an object. Property identifiers of type atom are only accepted, if it is a remote object (object implementation is only enforced if it is a local object). Numeric property identifiers are accepted regardless of remote object or not. For remote objects, this means you have to write "raw values" (usually Encoding structs).

ignore_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties get ignored, as if they were not specified.

revision - The BACnet protocol revision to check required properties against. Optional properties are regardless of revision available. See BACnet.Protocol.Constants.protocol_revision/0 for the available revisions.

skip_property_validation_remote_object - Skips property validation for remote objects. Sometimes it is possible that the value is invalid as per BACnet specification (i.e. value 0 for a multistate object), but you still want those to be represented. Value true neither type nor value are validated. Value :value means the type is still validated and only the value validator is not run (if present). The property's validator_fun will also be skipped.

object_opts()

@type object_opts() :: {:intrinsic_reporting, boolean()} | common_object_opts()

Options accepted when creating or configuring an Event Log object.

In addition to the common options, Event Log supports:

intrinsic_reporting - Enables intrinsic event reporting properties for the log.

property_name()

@type property_name() ::
  :acked_transitions
  | :buffer_size
  | :description
  | :enable
  | :event_algorithm_inhibit
  | :event_algorithm_inhibit_ref
  | :event_detection_enable
  | :event_enable
  | :event_message_texts
  | :event_message_texts_config
  | :event_state
  | :event_timestamps
  | :last_notify_record
  | :limit_enable
  | :log_buffer
  | :notification_class
  | :notification_threshold
  | :notify_type
  | :object_instance
  | :object_name
  | :profile_location
  | :profile_name
  | :record_count
  | :records_since_notification
  | :reliability
  | :reliability_evaluation_inhibit
  | :start_time
  | :status_flags
  | :stop_time
  | :stop_when_full
  | :tags
  | :time_delay
  | :time_delay_normal
  | :total_record_count

Available property names for this object.

property_update_error()

@type property_update_error() ::
  {:error,
   {error :: atom(),
    property :: BACnet.Protocol.Constants.property_identifier()}}

The structure for property errors.

t()

@type t() :: %BACnet.Protocol.ObjectTypes.EventLog{
  _metadata: internal_metadata(),
  _unknown_properties: %{
    optional(atom() | non_neg_integer()) =>
      term()
      | BACnet.Protocol.ApplicationTags.Encoding.t()
      | [BACnet.Protocol.ApplicationTags.Encoding.t()]
  },
  acked_transitions: BACnet.Protocol.EventTransitionBits.t() | nil,
  buffer_size: BACnet.Protocol.ApplicationTags.unsigned32(),
  description: String.t() | nil,
  enable: boolean(),
  event_algorithm_inhibit: boolean() | nil,
  event_algorithm_inhibit_ref: BACnet.Protocol.ObjectPropertyRef.t() | nil,
  event_detection_enable: boolean() | nil,
  event_enable: BACnet.Protocol.EventTransitionBits.t() | nil,
  event_message_texts: BACnet.Protocol.EventMessageTexts.t() | nil,
  event_message_texts_config: BACnet.Protocol.EventMessageTexts.t() | nil,
  event_state: BACnet.Protocol.Constants.event_state(),
  event_timestamps: BACnet.Protocol.EventTimestamps.t() | nil,
  last_notify_record: BACnet.Protocol.ApplicationTags.unsigned32() | nil,
  limit_enable: BACnet.Protocol.LimitEnable.t() | nil,
  log_buffer: [BACnet.Protocol.EventLogRecord.t()],
  notification_class: non_neg_integer() | nil,
  notification_threshold: BACnet.Protocol.ApplicationTags.unsigned32() | nil,
  notify_type: BACnet.Protocol.Constants.notify_type() | nil,
  object_instance: non_neg_integer(),
  object_name: String.t(),
  profile_location: String.t() | nil,
  profile_name: String.t() | nil,
  record_count: BACnet.Protocol.ApplicationTags.unsigned32(),
  records_since_notification:
    BACnet.Protocol.ApplicationTags.unsigned32() | nil,
  reliability: BACnet.Protocol.Constants.reliability() | nil,
  reliability_evaluation_inhibit: boolean() | nil,
  start_time: BACnet.Protocol.BACnetDateTime.t() | nil,
  status_flags: BACnet.Protocol.StatusFlags.t(),
  stop_time: BACnet.Protocol.BACnetDateTime.t() | nil,
  stop_when_full: boolean(),
  tags: BACnet.Protocol.BACnetArray.t(BACnet.Protocol.NameValue.t()) | nil,
  time_delay: non_neg_integer() | nil,
  time_delay_normal: non_neg_integer() | nil,
  total_record_count: BACnet.Protocol.ApplicationTags.unsigned32()
}

Represents a Event Log object. All keys should be treated as read-only, all updates should go only through update_property/3.

Properties which are for Intrinsic Reporting are nil, if disabled. If Intrinsic Reporting is enabled on the object, then the properties can not be nil.

Functions

add_property(object, property, value)

@spec add_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Adds an optional property to an object. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically added and instead the object must be newly created using create/4.

create(instance_number, object_name, properties \\ %{}, opts \\ [])

@spec create(
  non_neg_integer(),
  String.t(),
  %{optional(property_name() | atom() | non_neg_integer()) => term()},
  [object_opts() | internal_metadata()]
) :: {:ok, t()} | property_update_error()

Creates a new object struct with the defined properties. Optional properties are not created when not given, only required, given and dependency properties are created. Properties with a value of nil are ignored.

Only properties that are required for specific services (i.e. Intrinsic Reporting) are automatically created.

get_all_properties()

@spec get_all_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of all properties this object supports.

get_annotation(name)

@spec get_annotation(property_name()) :: [term()]

Auto generated function to get the annotations for the given property name.

get_annotations()

@spec get_annotations() :: [{name :: property_name(), values :: [term()]}]

Auto generated function to get the list of annotations for each property.

get_cov_properties()

@spec get_cov_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of properties used for COV reporting.

get_intrinsic_properties()

@spec get_intrinsic_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of intrinsic properties.

get_object_identifier(object)

@spec get_object_identifier(t()) :: BACnet.Protocol.ObjectIdentifier.t()

Get the BACnet object identifier.

get_optional_properties()

@spec get_optional_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of optional properties.

get_properties(object)

@spec get_properties(t()) :: [BACnet.Protocol.Constants.property_identifier()]

Get the list of properties the object has.

get_properties_type_map()

@spec get_properties_type_map() :: map()

Auto generated function to get a map of property name to type.

get_property(object, property)

@spec get_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, term()} | property_update_error()

Get a property's value from an object.

get_protected_properties()

@spec get_protected_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of protected properties.

Protected is an annotation and the object modules prevent writing to this property directly in code. The protected properties are either written on creation or updated automatically depending on other properties being written to. Some properties are only written once at creation and never updated.

get_readonly_properties()

@spec get_readonly_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of readonly properties.

Readonly is only an annotation that the property should be write protected on the BACnet side, there is no actual write protection in the object. This is a hint to the device server. If you need actual write protection, see protected.

get_required_properties()

@spec get_required_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of required properties.

has_property?(object, property)

@spec has_property?(t(), BACnet.Protocol.Constants.property_identifier()) :: boolean()

Checks if the given object has the given property.

See BACnet.Protocol.ObjectsUtility.has_property?/2 for implementation details.

intrinsic_reporting?(object)

@spec intrinsic_reporting?(t()) :: boolean()

Checks if the given object has Intrinsic Reporting enabled.

property_writable?(object, property)

@spec property_writable?(t(), BACnet.Protocol.Constants.property_identifier()) ::
  boolean()

Checks if the given property is writable.

Check BACnet.Protocol.ObjectsUtility.property_writable?/2 for a basic run-down.

remove_property(object, property)

@spec remove_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, t()} | property_update_error()

Removes an optional property from an object. This function is idempotent. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically removed and instead the object must be newly created using create/4. Required properties can not be removed.

supports_intrinsic()

@spec supports_intrinsic() :: boolean()

Auto generated function to check whether the object type supports intrinsic reporting.

update_property(object, property, value)

@spec update_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Updates a property of an object.