BACnet.Protocol.ObjectTypes.NotificationClass (bacstack v0.1.0)

The Notification Class object is the central configuration point for event/alarm routing in a BACnet device. Every object that can generate events (intrinsic or via Event Enrollment) references exactly one Notification Class. The class defines:

per-transition priorities (to-offnormal, to-fault, to-normal)

whether each transition requires operator acknowledgment

the list of recipients (who should receive the notification) together with optional time windows and process identifiers.

Because many objects can share the same Notification Class, a site can enforce consistent alarming policy (e.g. "all critical alarms go to the BMS at priority 8 and require ack") without repeating the recipient list in every object.

Object Description (ASHRAE 135)

The Notification Class object type defines a standardized object that represents and contains information required for the distribution of event notifications within BACnet systems.

Behaviour and Operation

Notification Class objects are pure configuration objects that control how and to whom event notifications are sent. Multiple event-generating objects (intrinsic or Event Enrollments) reference the same Notification Class to get consistent routing, priorities, and acknowledgment requirements.

The recipient_list (list of BACnet.Protocol.Destination entries) tells the notification engine which devices should receive Confirmed or Unconfirmed notifications for transitions, within optional time windows and with a given process identifier.

priority and ack_required structs give per-transition settings. The local notification engine (part of the device server or a central event router) is responsible for consulting the class when an event transition occurs and for managing acknowledgment state (acked_transitions lives on the event source, not here).

Developer Implementation Notes (geared to device server / application authors)

The generated code handles storage + basic mechanics (validation, implicit_relationships, readonly annotations as hints to your server, etc.). You must drive "special" live properties and side effects yourself, analogous to maintaining present_value on inputs via update_property/3 (never direct mutation). Read notes below + generated tables for details.

Special / live properties and expected developer behaviour

notification_class: Unique per class; used by event sources to reference which class handles their notifications.

priority: Per-transition (to-offnormal, to-fault, to-normal) priorities for the notifications.

ack_required: Which transitions require acknowledge by an operator (device).

recipient_list: Who gets notified (with filters for transitions, days, times, etc). Your notification router, when an event occurs for a source referencing this class, must evaluate the recipient_list (considering current time) and send the notifications (confirmed or unconfirmed) to the matching recipients. This is the "live" delivery list.

The object is pure config for the notification engine. The engine (part of your server) is responsible for consulting it on event transitions and delivering. acked_transitions live on the event source objects (Enrollment or intrinsic sources), not here.

Examples

Creating a Notification Class:

iex> {:ok, nc} = BACnet.Protocol.ObjectTypes.NotificationClass.create(700, "Alarms", %{notification_class: 700}); nc.object_name
"Alarms"

See Also

BACnet.Protocol.Destination

BACnet.Protocol.EventTransitionBits

BACnet.Protocol.NotificationClassPriority

The following part has been automatically generated.

Click to expand
This module defines a BACnet object of the type `notification_class`. The following properties are defined: | Property | Revision | Required | Readonly | Protected | Intrinsic | |----------|----------|----------|----------|-----------|-----------| | ack_required | | X | | | | | description | | | | | | | notification_class | | X | | | | | object_instance | | X | X | | | | object_name | | X | X | | | | priority | | X | | | | | profile_name | | | | | | | recipient_list | | X | | | | The following properties have additional semantics: | Property | Has Default | Has Init | Implicit Relationships | Validators | Annotations | |----------|-------------|----------|------------------------|------------|-------------| | ack_required | X | | | | | | priority | X | | | | | | recipient_list | X | | | | | The following table shows the default values and/or init functions: | Property | Default Value | Init Function | |----------|---------------|---------------| | ack_required | `%BACnet.Protocol.EventTransitionBits{...}` | | | priority | `%BACnet.Protocol.NotificationClassPriority{...}` | | | recipient_list | `[]` | |

Types

common_object_opts()

@type common_object_opts() ::
  {:allow_unknown_properties, boolean()}
  | {:ignore_unknown_properties, boolean()}
  | {:revision, BACnet.Protocol.Constants.protocol_revision()}

Common object options for creation - all are optional.

allow_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties (numeric identifiers usually means we dont know them) are accepted and put into a separate map. This does mean we can not validate or write them. Types of the values can be anything at this point. While you can read unknown properties with atom or integer as property identifier, you can only remove numeric unknown property identifiers from an object. Property identifiers of type atom are only accepted, if it is a remote object (object implementation is only enforced if it is a local object). Numeric property identifiers are accepted regardless of remote object or not. For remote objects, this means you have to write "raw values" (usually Encoding structs).

ignore_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties get ignored, as if they were not specified.

revision - The BACnet protocol revision to check required properties against. Optional properties are regardless of revision available. See BACnet.Protocol.Constants.protocol_revision/0 for the available revisions.

object_opts()

@type object_opts() :: common_object_opts()

Options accepted when creating or configuring a Notification Class object.

property_name()

@type property_name() ::
  :ack_required
  | :description
  | :notification_class
  | :object_instance
  | :object_name
  | :priority
  | :profile_name
  | :recipient_list

Available property names for this object.

property_update_error()

@type property_update_error() ::
  {:error,
   {error :: atom(),
    property :: BACnet.Protocol.Constants.property_identifier()}}

The structure for property errors.

t()

@type t() :: %BACnet.Protocol.ObjectTypes.NotificationClass{
  _metadata: internal_metadata(),
  _unknown_properties: %{
    optional(atom() | non_neg_integer()) =>
      term()
      | BACnet.Protocol.ApplicationTags.Encoding.t()
      | [BACnet.Protocol.ApplicationTags.Encoding.t()]
  },
  ack_required: BACnet.Protocol.EventTransitionBits.t(),
  description: String.t() | nil,
  notification_class: non_neg_integer(),
  object_instance: non_neg_integer(),
  object_name: String.t(),
  priority: BACnet.Protocol.NotificationClassPriority.t(),
  profile_name: String.t() | nil,
  recipient_list: [BACnet.Protocol.Destination.t()]
}

Represents a Notification Class object. All keys should be treated as read-only, all updates should go only through update_property/3.

Functions

add_property(object, property, value)

@spec add_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Adds an optional property to an object. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically added and instead the object must be newly created using create/4.

create(instance_number, object_name, properties \\ %{}, opts \\ [])

@spec create(
  non_neg_integer(),
  String.t(),
  %{optional(property_name() | atom() | non_neg_integer()) => term()},
  [object_opts() | internal_metadata()]
) :: {:ok, t()} | property_update_error()

Creates a new object struct with the defined properties. Optional properties are not created when not given, only required, given and dependency properties are created. Properties with a value of nil are ignored.

Only properties that are required for specific services (i.e. Intrinsic Reporting) are automatically created.

get_all_properties()

@spec get_all_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of all properties this object supports.

get_annotation(name)

@spec get_annotation(property_name()) :: [term()]

Auto generated function to get the annotations for the given property name.

get_annotations()

@spec get_annotations() :: [{name :: property_name(), values :: [term()]}]

Auto generated function to get the list of annotations for each property.

get_cov_properties()

@spec get_cov_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of properties used for COV reporting.

get_intrinsic_properties()

@spec get_intrinsic_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of intrinsic properties.

get_object_identifier(object)

@spec get_object_identifier(t()) :: BACnet.Protocol.ObjectIdentifier.t()

Get the BACnet object identifier.

get_optional_properties()

@spec get_optional_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of optional properties.

get_properties(object)

@spec get_properties(t()) :: [BACnet.Protocol.Constants.property_identifier()]

Get the list of properties the object has.

get_properties_type_map()

@spec get_properties_type_map() :: map()

Auto generated function to get a map of property name to type.

get_property(object, property)

@spec get_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, term()} | property_update_error()

Get a property's value from an object.

get_protected_properties()

@spec get_protected_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of protected properties.

Protected is an annotation and the object modules prevent writing to this property directly in code. The protected properties are either written on creation or updated automatically depending on other properties being written to. Some properties are only written once at creation and never updated.

get_readonly_properties()

@spec get_readonly_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of readonly properties.

Readonly is only an annotation that the property should be write protected on the BACnet side, there is no actual write protection in the object. This is a hint to the device server. If you need actual write protection, see protected.

get_required_properties()

@spec get_required_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of required properties.

has_property?(object, property)

@spec has_property?(t(), BACnet.Protocol.Constants.property_identifier()) :: boolean()

Checks if the given object has the given property.

See BACnet.Protocol.ObjectsUtility.has_property?/2 for implementation details.

property_writable?(object, property)

@spec property_writable?(t(), BACnet.Protocol.Constants.property_identifier()) ::
  boolean()

Checks if the given property is writable.

Check BACnet.Protocol.ObjectsUtility.property_writable?/2 for a basic run-down.

remove_property(object, property)

@spec remove_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, t()} | property_update_error()

Removes an optional property from an object. This function is idempotent. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically removed and instead the object must be newly created using create/4. Required properties can not be removed.

supports_intrinsic()

@spec supports_intrinsic() :: boolean()

Auto generated function to check whether the object type supports intrinsic reporting.

update_property(object, property, value)

@spec update_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Updates a property of an object.