BACnet.Protocol.ObjectTypes.StructuredView (bacstack v0.1.0)

The Structured View object is a pure organisational container. It lets a device expose a hierarchical tree (or forest) of references to other objects so that a client can navigate the device content the same way a user would navigate a building ("Floor 3" -> "AHU-03" -> "Supply Temp"). Each view contains a node_type, a node_subtype, and a list of subordinate_list entries that point to child objects (which may themselves be other Structured Views).

The structure is completely independent of the flat object_list in the Device object; it is an additional navigation aid. Views can represent physical hierarchy, functional systems, graphics pages, or any other grouping the implementer desires.

Object Description (ASHRAE 135)

The Structured View object type defines a standardized object that provides a container to hold references to subordinate objects, which may include other Structured View objects, thereby allowing multilevel hierarchies to be created.

Behaviour and Operation

Structured View objects are passive organisational containers. They do not contain or duplicate data from the referenced objects; each entry in subordinate_list is a reference.

The device or a configuration tool populates the tree. Clients use the view for navigation, graphics binding, or hierarchical browsing instead of (or in addition to) the flat object_list on the Device object. The view may reference objects that are not in the device's own object list (for proxy / remote views).

No active behaviour, no commandability, no event generation.

Developer Implementation Notes (geared to device server / application authors)

The generated code handles storage + basic mechanics (validation, implicit_relationships, readonly annotations as hints to your server, etc.). Read notes below + generated tables for details.

The "live" aspect is entirely the tree maintenance you perform as your object population changes. See "You build and maintain the tree" and "Typical uses" below.

A Structured View is a navigation aid, not a data container. It lets a client walk a tree (or DAG) that you have defined on top of your flat object population.

You build and maintain the tree: At device startup, after a configuration change, or when a new object is created, your code must ensure that the subordinate_list (and the node_type / node_subtype) on the relevant Structured View objects are correct. Each subordinate entry is just a reference; the real data lives in the target object.

Typical uses:

Physical hierarchy: Site -> Building -> Floor -> Room -> AHU -> Points

Functional systems: "Chilled Water System" -> all the pumps, valves, sensors that belong to it (even if they are scattered across several controllers)

Graphics pages: one view per graphic that lists exactly the points the page needs to bind

Proxy / remote views: a view on device A that contains references to objects that actually live on device B (the client follows the references with normal ReadProperty calls)

No automatic maintenance: If you delete an object that is referenced from several views, you are responsible for cleaning up the subordinate_list entries. The Structured View object does not know about object deletion.

Writing the view: Because the lists are normal properties, a sufficiently privileged client (or another tool) can completely reorganise your hierarchy at runtime.

Performance for clients: A client that wants to draw a whole building can start at a root Structured View and recursively follow the subordinate references. Because each reference can be to a local or remote object, the client will issue the appropriate ReadProperty calls.

Structured Views are the main mechanism a BACnet device has to impose structure on what would otherwise be a flat bag of objects.

Examples

Creating a Structured View:

iex> {:ok, sv} = BACnet.Protocol.ObjectTypes.StructuredView.create(1700, "BuildingView", %{node_type: :device}); sv.object_name
"BuildingView"

The following part has been automatically generated.

Click to expand
This module defines a BACnet object of the type `structured_view`. The following properties are defined: | Property | Revision | Required | Readonly | Protected | Intrinsic | |----------|----------|----------|----------|-----------|-----------| | description | | | | | | | node_subtype | | | | | | | node_type | | X | | | | | object_instance | | X | X | | | | object_name | | X | X | | | | profile_location | 19 | | | | | | profile_name | | | | | | | subordinate_annotations | | | | | | | subordinate_list | | X | | | | | tags | 19 | | | | | The following properties have additional semantics: | Property | Has Default | Has Init | Implicit Relationships | Validators | Annotations | |----------|-------------|----------|------------------------|------------|-------------| | profile_location | | | | Fun | `revision: 19` | | subordinate_list | X | | | | | | tags | | | | | `revision: 19` | The following table shows the default values and/or init functions: | Property | Default Value | Init Function | |----------|---------------|---------------| | subordinate_list | `[]` | |

Types

common_object_opts()

@type common_object_opts() ::
  {:allow_unknown_properties, boolean()}
  | {:ignore_unknown_properties, boolean()}
  | {:revision, BACnet.Protocol.Constants.protocol_revision()}
  | {:skip_property_validation_remote_object, boolean() | :value}

Common object options for creation - all are optional.

allow_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties (numeric identifiers usually means we dont know them) are accepted and put into a separate map. This does mean we can not validate or write them. Types of the values can be anything at this point. While you can read unknown properties with atom or integer as property identifier, you can only remove numeric unknown property identifiers from an object. Property identifiers of type atom are only accepted, if it is a remote object (object implementation is only enforced if it is a local object). Numeric property identifiers are accepted regardless of remote object or not. For remote objects, this means you have to write "raw values" (usually Encoding structs).

ignore_unknown_properties - Properties that are unknown to the object implementation are usually rejected. With this option, unknown properties get ignored, as if they were not specified.

revision - The BACnet protocol revision to check required properties against. Optional properties are regardless of revision available. See BACnet.Protocol.Constants.protocol_revision/0 for the available revisions.

skip_property_validation_remote_object - Skips property validation for remote objects. Sometimes it is possible that the value is invalid as per BACnet specification (i.e. value 0 for a multistate object), but you still want those to be represented. Value true neither type nor value are validated. Value :value means the type is still validated and only the value validator is not run (if present). The property's validator_fun will also be skipped.

object_opts()

@type object_opts() :: common_object_opts()

Options accepted when creating or configuring a Structured View object.

property_name()

@type property_name() ::
  :description
  | :node_subtype
  | :node_type
  | :object_instance
  | :object_name
  | :profile_location
  | :profile_name
  | :subordinate_annotations
  | :subordinate_list
  | :tags

Available property names for this object.

property_update_error()

@type property_update_error() ::
  {:error,
   {error :: atom(),
    property :: BACnet.Protocol.Constants.property_identifier()}}

The structure for property errors.

t()

@type t() :: %BACnet.Protocol.ObjectTypes.StructuredView{
  _metadata: internal_metadata(),
  _unknown_properties: %{
    optional(atom() | non_neg_integer()) =>
      term()
      | BACnet.Protocol.ApplicationTags.Encoding.t()
      | [BACnet.Protocol.ApplicationTags.Encoding.t()]
  },
  description: String.t() | nil,
  node_subtype: String.t() | nil,
  node_type: BACnet.Protocol.Constants.node_type(),
  object_instance: non_neg_integer(),
  object_name: String.t(),
  profile_location: String.t() | nil,
  profile_name: String.t() | nil,
  subordinate_annotations: [String.t()] | nil,
  subordinate_list: [BACnet.Protocol.DeviceObjectRef.t()],
  tags: BACnet.Protocol.BACnetArray.t(BACnet.Protocol.NameValue.t()) | nil
}

Represents a Structured View object. All keys should be treated as read-only, all updates should go only through update_property/3.

Functions

add_property(object, property, value)

@spec add_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Adds an optional property to an object. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically added and instead the object must be newly created using create/4.

create(instance_number, object_name, properties \\ %{}, opts \\ [])

@spec create(
  non_neg_integer(),
  String.t(),
  %{optional(property_name() | atom() | non_neg_integer()) => term()},
  [object_opts() | internal_metadata()]
) :: {:ok, t()} | property_update_error()

Creates a new object struct with the defined properties. Optional properties are not created when not given, only required, given and dependency properties are created. Properties with a value of nil are ignored.

Only properties that are required for specific services (i.e. Intrinsic Reporting) are automatically created.

get_all_properties()

@spec get_all_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of all properties this object supports.

get_annotation(name)

@spec get_annotation(property_name()) :: [term()]

Auto generated function to get the annotations for the given property name.

get_annotations()

@spec get_annotations() :: [{name :: property_name(), values :: [term()]}]

Auto generated function to get the list of annotations for each property.

get_cov_properties()

@spec get_cov_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of properties used for COV reporting.

get_intrinsic_properties()

@spec get_intrinsic_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of intrinsic properties.

get_object_identifier(object)

@spec get_object_identifier(t()) :: BACnet.Protocol.ObjectIdentifier.t()

Get the BACnet object identifier.

get_optional_properties()

@spec get_optional_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of optional properties.

get_properties(object)

@spec get_properties(t()) :: [BACnet.Protocol.Constants.property_identifier()]

Get the list of properties the object has.

get_properties_type_map()

@spec get_properties_type_map() :: map()

Auto generated function to get a map of property name to type.

get_property(object, property)

@spec get_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, term()} | property_update_error()

Get a property's value from an object.

get_protected_properties()

@spec get_protected_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of protected properties.

Protected is an annotation and the object modules prevent writing to this property directly in code. The protected properties are either written on creation or updated automatically depending on other properties being written to. Some properties are only written once at creation and never updated.

get_readonly_properties()

@spec get_readonly_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of readonly properties.

Readonly is only an annotation that the property should be write protected on the BACnet side, there is no actual write protection in the object. This is a hint to the device server. If you need actual write protection, see protected.

get_required_properties()

@spec get_required_properties() :: [BACnet.Protocol.Constants.property_identifier()]

Auto generated function to get the names of required properties.

has_property?(object, property)

@spec has_property?(t(), BACnet.Protocol.Constants.property_identifier()) :: boolean()

Checks if the given object has the given property.

See BACnet.Protocol.ObjectsUtility.has_property?/2 for implementation details.

property_writable?(object, property)

@spec property_writable?(t(), BACnet.Protocol.Constants.property_identifier()) ::
  boolean()

Checks if the given property is writable.

Check BACnet.Protocol.ObjectsUtility.property_writable?/2 for a basic run-down.

remove_property(object, property)

@spec remove_property(
  t(),
  BACnet.Protocol.Constants.property_identifier() | non_neg_integer()
) ::
  {:ok, t()} | property_update_error()

Removes an optional property from an object. This function is idempotent. Remote objects can not be mutated using this operation.

Please note that properties of services can not be dynamically removed and instead the object must be newly created using create/4. Required properties can not be removed.

supports_intrinsic()

@spec supports_intrinsic() :: boolean()

Auto generated function to check whether the object type supports intrinsic reporting.

update_property(object, property, value)

@spec update_property(t(), BACnet.Protocol.Constants.property_identifier(), term()) ::
  {:ok, t()} | property_update_error()

Updates a property of an object.