BACnet.Stack.Client (bacstack v0.1.0)

The BACnet client is responsible for connecting the application to the BACnet transport protocol and vice versa - it interfaces with the BACnet transport protocol, using the transport behaviour. The client will take requests and send them to the BACnet transport protocol and ultimately listen for frames from the BACnet transport protocol.

The application will receive process messages on the specified notification_receiver. If there's no notification receiver and a confirmed service request is received, the client will automatically send an Reject APDU with reason :other to the remote BACnet device, the application won't be informed in any way.

BACnet BVLL and BACnet NPDU are directly forwarded to the application, without any processing.

For BACnet BVLL, the following message is sent:

{:bacnet_transport, protocol_id, source_address, {:bvlc, bvlc}, portal}

This is the same as BACnet.Stack.TransportBehaviour.transport_msg/0.

For BACnet NDPU, the following message is sent:

{:bacnet_transport, protocol_id, source_address, {:network, bvlc, npci, nsdu}, portal}

This is the same as BACnet.Stack.TransportBehaviour.transport_msg/0.

For BACnet APDU, the following message is sent:

{:bacnet_client, reference() | nil, apdu, {source_address, bvlc, npci}, pid()}

The reference is only present on confirmed service requests and is used for reply/4. APDU is BACnet.Protocol.apdu/0. The PID is of the Client process. For the rest, see BACnet.Stack.TransportBehaviour.transport_msg/0.

BACnet APDUs are decoded, checked against the internal cache and forwarded to the application. The client has an internal cache for requests for the application and replies from the application to the transport protocol in order to deduplicate request (i.e. when the application takes for the remote BACnet device far too long and re-sends the APDU). In case of duplicated request, the request will not be forwarded to the application, instead if a response is available in the cache, the response will be sent to the transport (i.e. the case when the response does not arrive at the remote BACnet device). Individual APDU timeouts on a per-source basis can be applied using set_apdu_timeouts/2.

The client will also keep track of sent APDUs for confirmed service requests and re-send them automatically, if no response arrives in the APDU timeout timeframe. If the maximum APDU retries is reached, the request will be deleted and the application will get an APDU timeout response.

If the application takes too long to respond to a remote BACnet request, the client will automatically send an Abort APDU and respond to a reply request from the application with an app timeout response.

If the application replies to a routed request (Forwarded NPDU), the client will automatically set the correct destination address in reply/4.

The BACnet client will not automatically respond to Who-Is, Who-Has, Time-Synchronization, etc. queries, as this is outside of the responsibility of this low level BACnet client implementation.

Types

apdu_timeouts()

@type apdu_timeouts() :: %{
  optional({source_address :: term(), device_id :: non_neg_integer() | nil}) =>
    apdu_timeout :: non_neg_integer()
}

Per-source APDU timeouts.

Device ID is only known, if the source transmits it with the APDU, as such, most of the time it can be nil.

start_option()

@type start_option() ::
  {:apdu_retries, pos_integer()}
  | {:apdu_timeout, pos_integer()}
  | {:disable_app_timeout, boolean()}
  | {:disable_resp_cache, boolean()}
  | {:notification_receiver, Process.dest() | [Process.dest()]}
  | {:npci_source, BACnet.Protocol.NpciTarget.t()}
  | {:segmentator, GenServer.server()}
  | {:segments_store, GenServer.server()}
  | {:segmented_rcv_window_overwrite, boolean()}
  | {:transport,
     module() | {module(), BACnet.Stack.TransportBehaviour.transport()}}
  | GenServer.option()

Valid start options. For a description of each, see start_link/1.

start_options()

@type start_options() :: [start_option()]

List of start options.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get_apdu_timeouts(server)

@spec get_apdu_timeouts(GenServer.server()) :: {:ok, apdu_timeouts()}

Get the per-source APDU timeouts map. This is only used for receiving.

Each source is identified by source address and device ID (device ID is only known if the source transmit it in the BACnet NPCI).

reply(server, ref, reply, opts \\ [])

@spec reply(GenServer.server(), reference(), BACnet.Protocol.apdu(), Keyword.t()) ::
  :ok
  | {:error, :app_timeout}
  | {:error, term()}
  | {:error, {Exception.t(), stacktrace :: Exception.stacktrace()}}

Replies to a confirmed service request from a remote BACnet device. The APDU frame may be segmentated by the client, depending on the APDU size and maximum transmittable APDU size.

The reference identifies the request in the client. The request is hold in the client to be able to apply application timeout constraints and automatically re-send responses on transmission errors (remote BACnet device sends the same request again). The reference is given as part of the BACnet client notification process message.

If an automatic application timeout response has been sent (Abort APDU), {:error, :app_timeout} will be returned when trying to reply to the request.

See send/4 for more information about opts. Both :max_segments and :segmentation_supported of opts are automatically derived from the confirmed service request, if not explicitely given.

send(server, destination, data, opts \\ [])

@spec send(GenServer.server(), term(), BACnet.Protocol.apdu(), Keyword.t()) ::
  :ok
  | {:ok, BACnet.Protocol.apdu()}
  | {:error, :apdu_timeout}
  | {:error, term()}
  | {:error, {Exception.t(), stacktrace :: Exception.stacktrace()}}

Sends the given APDU frame to the specified destination (remote BACnet device). The APDU frame may be segmentated by the client.

The client will keep track of sent confirmed service requests and automatically re-send the APDUs, if the APDU times out, without a response from the remote BACnet server. If the maximum APDU retry count gets reached, {:error, :apdu_timeout} will be returned.

This function returns, for confirmed service requests, after receiving the response from the remote BACnet server, for everything else almost immediately.

As such, this function will block for maximum 60000ms, before the backpressure mechanism will exit the caller.

Destination depends on the transport module and is validated against the transport module.

BACnet Abort/Error/Reject are also returned in :ok tuples, not only acknowledgements and requests.

When sending and the APDU is too large and thus is needed to be segmented, the client will check accordinging to the given options, whether segmentation can occur and how many segments are supported. If the remote device does not support segmentation or a buffer overflow would occur due to too many segments, this client will send an Abort APDU to the remote device and return an error to the caller. The same will happen for too long APDUs that can not be segmented.

See the BACnet.Stack.TransportBehaviour.send/4 documentation for what opts can be. In addition the following are available:

max_apdu_length: pos_integer() - Optional. The maximum APDU length the remote BACnet device supports (defaults to the transport max APDU length).

max_segments: pos_integer() - Optional. Maximum amount of segments the remote BACnet device can accept (defaults to 2).

segmentation_supported: Constants.segmentation() - Optional. Which kind of segmentation the remote BACnet device supports (defaults to none).

set_apdu_timeouts(server, timeouts)

@spec set_apdu_timeouts(GenServer.server(), apdu_timeouts()) :: :ok

Set the per-source APDU timeouts map. This is only used for receiving.

Each source is identified by source address and device ID (device ID is only known if the source transmit it in the BACnet NPCI).

start_link(opts)

@spec start_link(start_options()) :: GenServer.on_start()

Starts and links the BACnet Client.

The following options are available, in addition to GenServer.options/0:

apdu_retries: pos_integer() - Optional. The amount of APDU sending retries (defaults to 3). Only applied to confirmed service requests.

apdu_timeout: pos_integer() - Optional. The APDU timeout to be waiting for a response, in ms (defaults to 3000ms). Only applied to confirmed service requests.

disable_app_timeout: boolean() - Optional. Disables the application timeout mechanism.

disable_resp_cache: boolean() - Optional. Disables the APDU response cache and deduplication (see also reply/4).

notification_receiver: Process.dest() | [Process.dest()] - Optional. The destination to send messages to.

npci_source: NpciTarget.t() - Optional. The NPCI target to use as source for outgoing APDUs.

segmentator: GenServer.server() - Required. The segmentator server to use.

segments_store: GenServer.server() - Required. The segments store server to use.

segmented_rcv_window_overwrite: boolean() - Optional. Enable to overwrite the window size to 1 for incoming segmented APDUs when it is bound to be routed (i.e. subject to BACnet/IP UDP packet re-ordering). If you're having difficulty receiving segmented APDUs and the packets get routed on BACnet/IP, you should consider enabling this and see if it helps.

transport: module() | {module(), TransportBehaviour.transport()} - Required. The transport to use. module() is equivalent to {module(), module()} (the module name is registered process name). The given transport must implement the TransportBehaviour behaviour.