BACnet.Stack.LogBuffer (bacstack v0.1.0)

Simple log buffer implementation for event and trend logs. The type of the item is not enforced.

The log buffer can maintain a maximum buffer size by dropping the oldest item on overflow (ring buffer).

Types

item()

@type item() ::
  BACnet.Protocol.EventLogRecord.t()
  | BACnet.Protocol.LogRecord.t()
  | BACnet.Protocol.LogMultipleRecord.t()

Log buffer item type.

items()

@opaque items()

Implementation detail and thus private API. Changes to it do not count towards Semantic Versioning.

t()

@type t() :: %BACnet.Stack.LogBuffer{
  items: items(),
  max_size: pos_integer() | nil,
  size: non_neg_integer()
}

Representative type for the log buffer.

Functions

checkin(buffer, item)

@spec checkin(t(), item()) :: t()

Checks in an item.

If the log buffer has a max size, the max size will be maintained (meaning the oldest item gets dropped on overflow).

checkout(buffer)

@spec checkout(t()) :: {item :: item() | nil, new_buffer :: t()}

Checks out the oldest item in the log buffer.

If the log buffer is empty, nil will be returned as item.

drop(buffer, amount \\ 1)

@spec drop(t(), pos_integer()) :: t()

Drops the specified amount of oldest items from the log buffer.

from_list(list, max_size \\ nil)

@spec from_list(list(), pos_integer() | nil) :: t()

Creates a new log buffer from the given list, with an optional max size. The head of the list will be the oldest entry in the log buffer.

If a max size is specified, the list will be truncated to the max size (from the end). The queue length will be calculated in any case. This operation is O(n).

get_size(buffer)

@spec get_size(t()) :: non_neg_integer()

Get the size of the log buffer.

new(max_size \\ nil)

@spec new(pos_integer() | nil) :: t()

Creates a new log buffer with an optional max size.

peek(buffer)

@spec peek(t()) :: {:ok, item()} | :error

Peeks into the log buffer and returns the oldest item without removing it.

peek_r(buffer)

@spec peek_r(t()) :: {:ok, item()} | :error

Peeks into the log buffer and returns the newest item without removing it.

read_range(buffer, offset \\ 0, count)

@spec read_range(t(), non_neg_integer(), pos_integer()) :: [item()]

Reads the specified count, starting from the given position (offset), from the log buffer.

to_list(buffer)

@spec to_list(t()) :: [item()]

Creates a list from the log buffer.

truncate(buffer)

@spec truncate(t()) :: t()

Truncates the log buffer to size zero.