BACnet.Protocol.APDU (bacstack v0.1.0)

This module provides decoding of Application Data Units (APDU). Encoding of APDUs are directly handled in the APDU modules.

APDUs can be segmented and thus will require processing and merging the segments. The module BACnet.Stack.SegmentsStore fulfills this purpose and thus all incomplete tuples received from decode/1 should be passed to an instance of that module (preferably under a supervisor). Only ComplexACK and ConfirmedServiceRequest APDUs can be segmented, as specified by the BACnet protocol specification. See also the BACnet.Stack.SegmentsStore module documentation.

See also:

BACnet.Protocol.APDU.Abort

BACnet.Protocol.APDU.ComplexACK

BACnet.Protocol.APDU.ConfirmedServiceRequest

BACnet.Protocol.APDU.Error

BACnet.Protocol.APDU.Reject

BACnet.Protocol.APDU.SegmentACK

BACnet.Protocol.APDU.SimpleACK

BACnet.Protocol.APDU.UnconfirmedServiceRequest

Functions

decode(data)

@spec decode(binary()) ::
  {:ok, BACnet.Protocol.apdu()}
  | {:error, term()}
  | {:incomplete, BACnet.Protocol.IncompleteAPDU.t()}

Decodes the APDU. The binary data must contain only APDU data.

If the APDU is segmented, this function will return an incomplete tuple, which must be handled by the BACnet.Stack.SegmentsStore module.

See the BACnet.Stack.SegmentsStore module documentation for more information about incoming segmentation (:segmented_receive).

decode_abort(apdu_type, apdu_data)

@spec decode_abort(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.Abort.t()} | {:error, term()}

Decodes the Abort APDU from binary data.

decode_complex_ack(apdu_type, apdu_data)

@spec decode_complex_ack(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.ComplexACK.t()}
  | {:error, term()}
  | {:incomplete, BACnet.Protocol.IncompleteAPDU.t()}

Decodes the Complex ACK APDU from binary data.

When encountering segmentation, this function will return an incomplete tuple. See BACnet.Protocol module for more information.

decode_confirmed_request(apdu_type, apdu_data)

@spec decode_confirmed_request(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.ConfirmedServiceRequest.t()}
  | {:error, term()}
  | {:incomplete, BACnet.Protocol.IncompleteAPDU.t()}

Decodes the Confirmed Service Request APDU from binary data.

When encountering segmentation, this function will return an incomplete tuple. See BACnet.Protocol module for more information.

decode_error(apdu_type, apdu_data)

@spec decode_error(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.Error.t()} | {:error, term()}

Decodes the Error APDU from binary data.

decode_reject(apdu_type, apdu_data)

@spec decode_reject(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.Reject.t()} | {:error, term()}

Decodes the Reject APDU from binary data.

decode_segment_ack(apdu_type, apdu_data)

@spec decode_segment_ack(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.SegmentACK.t()} | {:error, term()}

Decodes the Segment ACK APDU from binary data.

decode_simple_ack(apdu_type, apdu_data)

@spec decode_simple_ack(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.SimpleACK.t()} | {:error, term()}

Decodes the Simple ACK APDU from binary data.

decode_unconfirmed_request(apdu_type, apdu_data)

@spec decode_unconfirmed_request(byte(), binary()) ::
  {:ok, BACnet.Protocol.APDU.UnconfirmedServiceRequest.t()} | {:error, term()}

Decodes the Unconfirmed Service Request APDU from binary data.

get_invoke_id_from_raw_apdu(apdu)

@spec get_invoke_id_from_raw_apdu(binary()) ::
  {:ok, invoke_id :: byte()} | {:error, term()}

Extracts the invoke ID from the given raw APDU.

This is useful for replying to APDUs, which can not be properly fully decoded.